Reland: Reorganize media::VideoCapture* types
Relanding after revert in r236927.  Original commit message below:

The purpose of this CL is to clean up the distinction between
VideoCaptureFormat (which identifies the captured type of a frame),
VideoCaptureParams (which identifies the requested format of a capture),
and VideoCaptureCapability (which identifies the capture capabilities of a
device).

Notably:
* VideoCaptureFormat::frame_size_type ->
  VideoCaptureParams::allow_resolution_change, as variable resolution
  capability is a per-session, not a per-frame property.
* VideoCaptureCapability::color -> VideoCaptureFormat::pixel_format, as frame
  color format is a per-frame property.
  * As VideoCaptureParams holds a VideoCaptureFormat member, capture requests
    are able now to request a particular capture color format.

BUG=269312
TEST=local build, run unittests, chrome on CrOS snow, desktop Linux
TBR=ncarter@chromium.org,jiayl@chromium.org,yzshen@chromium.org,wjia@chromium.org,jschuh@chromium.org

Review URL: https://codereview.chromium.org/83633008