wpad: do dhcp discovery before trying quickcheck
When we're deciding what state to transition into after a wait or a fetch
failure, consider whether the source we're about to try is DHCP or DNS. If it's
DHCP, skip QuickCheck, since QuickCheck is irrelevant for DHCP sources. Also,
add a new unit test covering this case, and a MockDhcpProxyScriptFetcher class
for testing.

TEST=unit
BUG=320694

Review URL: https://codereview.chromium.org/63373005