Fixed regression in about:policy from revision 235005.
The Profile's PolicyService does not include the DeviceCloudPolicyManager since
https://codereview.chromium.org/60823003, but about:policy reads policy data
from that service.

This change adds the DeviceCloudPolicyManager to the Profile's PolicyService
too, but unlike the platform provide it is not wrapped by a
ForwardingPolicyProvider, since it's not currently used as a source of
policy for components for the profile.

BUG=321281

Review URL: https://codereview.chromium.org/83413003