[NaCl SDK] fix shared library link rule in example build system.
The rules for building shared libraries are now much
more similar to those for building executables. In
particular we honor the libraries paths and linker flags
for the project.

This discrepancy was causing the libraries to be linked
without -pthread, which was causing a failure under the
recent binutils update in the NaCl toolchain.

R=binji@chromium.org, mcgrathr@chromium.org

Review URL: https://codereview.chromium.org/84073002