IndexedDB: Don't re-send WebIDBDatabase pointer
Originally, the WebIDBDatabase impl pointer was handed off from Chrome
to WebKit in an onSuccess() callback. When "upgradeneeded" support was
added, which fires earlier and needs to deliver the connection object
to script, the same WebIDBDatabase impl pointer was sent in both
onUpgradeNeeded() and onSuccess(). This violates the "pointer is a
transfer of ownership" contract with the API and required
extra plumbing on both sides of the API.

The blink public API has been cleaned up so this is no longer necessary;
don't re-send the pointer in onSuccess() if it was previously sent in
onUpgradeNeeded().

R=alecflett@chromium.org
BUG=234277

Review URL: https://codereview.chromium.org/86053002