Fullscreen app windows cannot have always-on-top enabled
An app window in fullscreen mode should not also be always-on-top. This
is because it makes it difficult to alt+tab to another window, which is
both annoying from a UX perspective and a security issue.

To mitigate this:
The always-on-top property is temporarily switched off when the window
enters fullscreen, but then reinstated when the window is restored.
To be consistent, if a window is fullscreen and an attempt is made to
set it always-on-top, we only apply the always-on-top property when the
window exits fullscreen.

This patch also changes the value returned by the chrome.app.window 
IsAlwaysOnTop() function. It previously returned the actual state of
the window, but now returns the desired state, i.e. the value will
be true even when the property is switched off in fullscreen mode.

BUG=320103, 171597
TESTS=browser_tests (PlatformAppBrowserTest.WindowsApiAlwaysOnTop*).
interactive_ui_tests (ShellWindowTest.*AlwaysOnTop*).
Also test manually using Window State chrome sample app using app 
window API fullscreen and HTML5 fullscreen.

Review URL: https://codereview.chromium.org/76993004