Fix windowless NPAPI Flash videos not working on certain sites.
This is a continuation of the previous patch https://codereview.chromium.org/69953030 which addressed this on
some sites. We intercepted the WindowFromPoint API in that patch and returned the handle to the dummy plugin window.

There was a bug in the WindowFromPoint patch which was returning the first available child window in the parent window
hierarchy. We need to invoke the ChildWindowFromPoint API to get the actual child window and return it if it is the dummy
window.

The other changes are as below:-
1. The RenderWidgetHostViewAura class was not destroying the dummy window when the instance was destroyed.

2. When the RenderWidgetHostViewAura class is hidden due to the user switching to another tab, we set the bounds of the hidden
   window to 0. When the RenderWidgetHostViewAura instance is shown we restore the bounds of the hidden window. This is to ensure
   that the ChildWindowFromPoint function used in the plugin process returns the correct HWND.

3. When the RenderWidgetHostViewAura instance is added to a root window we need to reparent the dummy plugin window to the
   new window. This magic happens in the RenderWidgetHostImpl::ParentChanged function. Previously we were reparenting the
   dummy windows in the plugin process. With the dummy plugin window in the browser hierarchy we only need to reparent this
   window to the new browser window.

4. We have a new method GetPluginViewId in the RenderWidgetHostViewPort interface for Aura Windows. This returns the dummy plugin
   parent window.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=301548

BUG=301548
R=cpu@chromium.org, jam@chromium.org, jam

Review URL: https://codereview.chromium.org/83633007