Release media player resources when device is locked
Stop holding on resources when the device gets locked while
fullscreen media player is active. This is the same as what
we do for non-fullscreen and the MediaPlayerAndroid implementations
alreay has logic to be able to restart after Release is called.
The special case for fullscreen has been introduced as a workaround
because restarting did not work in fullscreen. This cl removes
the special case and fix the original issue: which is that we drop
the reference to the surface we got from the fullscreen player so
when we restart we could not set up the new MediaPlayer to draw onto it.

BUG=322181

Review URL: https://codereview.chromium.org/81693003