Reapply DeviceDataRoamingEnabled policy on device list changes.
Previously, this policy was applied once upon policy load, failing silently if no cellular device was found. If the device became visible afterwards, it wasn't configured. As Roaming might be disabled and the policy prevents the user from enabling it, this broke roaming on enrolled devices.

Now, the policy is reapplied on each change of the network device list.

BUG=321934
R=pastarmovj@chromium.org

Review URL: https://codereview.chromium.org/88323002