net: fix PORT_Assert crash when renegotiating.
r235826 added a test when sending the ChannelID extension which is
nonsense in at least two different ways.

Firstly, it PORT_Asserts when it doesn't match, which is wrong because it's not
an exceptional condition. The handshake hash is only stored in the session if
ChannelID was negotiated so this should have been crashing on every,
non-ChannelID resume.

The reason that it didn't crash was because the isResuming flag that it used
wasn't set until the ServerHello was processed. So the only time when it did
crash was on renegotiation after a resumption.

This change fixes the resumption test and removes the assert. I think the
resumption test is needed because it's possible that a server could deploy
ChannelID between the initial handshake and a resumption attempt.

BUG=none

Review URL: https://codereview.chromium.org/85783004