IndexedDB: Don't leak database proxy objects
If a connection is after OnUpgradeNeeded it is removed from a map and
when the OnSuccess comes through a new WebIDBDatabase object is minted
but ignored by the blink side. This leads to a leak, which was turned
up by http://crrev.com/67463006 which starts asserting that a new
object is not in a success following an upgrade.

Push the change from r237142 back into the browser side of things, and
don't even send an ID across.

R=alecflett@chromium.org
TBR=darin@chromium.org
BUG=234277

Review URL: https://codereview.chromium.org/86773002