Ensure that the OSK on Windows 8 shows up when we tap on an editable field in any WebContents.
The current implementation was displaying the OSK only when we tapped on a WebContents hosted by a Browser window. This
relied on the ChromeRenderViewObserver object in the renderer sending an IPC message ChromeViewHostMsg_FocusedNodeTouched
which is handled by the Chrome browser code via a WebContentsObserver implementation.

Based on discussions with jam and comments from avi in the review, we decided to implement this functionality in content.

Fixes are as below:-
1. We have a new IPc message ViewHostMsg_FocusedNodeTouched which is defined in the view_messages.h file.
2. This IPC is sent by the RenderViewImpl::didHandleGestureEvent method if we tapped on an editable field.
3, This IPC is handled by the RenderViewHostImpl class. We display the OSK on Windows 8 Aura if a focused editable
   node was tapped.
4. The ChromeViewHostMsg_FocusedNodeTouched IPC has been deleted along with the didHandleGestureEvent method in the
   RenderViewObserver interface.

BUG=321576, 319219
R=jam@chromium.org, palmer@chromium.org, jam
TBR=jschuh

Review URL: https://codereview.chromium.org/89873002