[Mojo] Remove static "bootstrap" state in mojo_js
This CL removes the static static in boostrap.cc that we used to transfer the
initial handle from C++ into JavaScript. Rather that storing the initial handle
in static state and evaling some script to kick off the loading process, we now
use an API to load a module by name and then pass the initial handle to the
function defined by that module.

This CL paves the say for adding threading support to mojo_js. The previous
static approach didn't scale to multiple threads, but the approach in this CL
does. I'll actually add the API to create a thread in the next CL.

R=aa@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/90203002