Fix for GL renderer's GL scissor state being shared with Ganesh.
When we do GPU rasterization of tiles, the GL context is
shared between Ganesh and the compositor. Both systems
assume that the GL state won't change in between calls
which is not enforced at the moment. 

This patch fixes such a problem with the GL scissor rect.

R=enne

Review URL: https://codereview.chromium.org/86833002