Transition to HAVE_METADATA when seeking.
Updating WebMediaPlayerImpl to transition to HAVE_METADATA when a seek
is requested. This is the first in a series of changes that will allow
the actual buffering state to be properly communicated to JavaScript.
This change properly reflects that we are not ready to immediately play
when a seek is requested.

BUG=144683
TESTS=All existing LayoutTests still pass. New Blink LayoutTests will be
 landed once this change lands to verify that the new readyState
 transitions.

Review URL: https://codereview.chromium.org/90283002