More aggressively evict textures used by the parent compositor.
Do not put textures that are in use by the parent compositor
at the back of the eviction queue. Rather, put them in
according to their priority, and use being in the parent
compositor to break ties (preferring to evict resources in
the parent compositor, because only once they are evicted
is there a chance that they may be deleted).

Rename CanBeRecycled to CanBeRecycledIfNotInExternalUse.
This function was used in the sorting function, and
CanBeRecycled is false for textures that are in the
parent compositor. Can CanBeRecycled to not take into
account the parent compositor, and rename the function
accordingly.

BUG=322299

Review URL: https://codereview.chromium.org/88913007