Workaround for problem where our javascript loads slow.
This is a very speculative fix since it's hard to repro locally. If this
patch solves the problem, there's a race between js loading and the test
which likely requires a deep look at how the navigation logic and test
javascript helpers interact with each other.

TBR=tommi@chromium.org
BUG=281268

Review URL: https://codereview.chromium.org/94513002