Fix a problem with no video with chromecast
This patch is to fix a regression in M32 for chromecast.

When calling tabCapture() to create a MediaStream the bug caused a
video stream to not be created. This is due to the fast that device id
is empty for tab capture streams.

This led the code in MediaStreamImpl::FindLocalSource() to think that
the new video device is already created. The consequence is there's no
video device added and a duplicated audio stream is added to
PeerConnection.

The solution to this problem is simple: add device type to the check.
This should ensure that audio and video devices do not match.

BUG=324174

Review URL: https://codereview.chromium.org/93823002