Check library version and handle library load errors
This CL modifies the build to incorporate the expected C++ library version
in the Java code. This is then checked when the library is loaded, to
make sure that the C++ and Java builds match.

This CL also implements error handling when library loads fail or the
loaded version doesn't match the expected version.

BUG=311644

Review URL: https://codereview.chromium.org/59533009