Android: split golden content from jni_generator_tests.
Rather than intermix the generated golden files with the test script,
extract in separate files.
Also, if "REBASELINE" environment variable is set, it'll re-generate the
golden scripts, to it's much simpler to do changes in the generator itself.
Note: there are NO functional changes in the generator itself, only in the tests.

BUG=

Review URL: https://codereview.chromium.org/96523004