Set long-press affordance widget as non-focusable
The widget showing long-press affordance was focusable, but since it was
inserted in Ash's OverlayContainer layer which was non-focusable, the
affordance was ultimately non-focusable. After a recent change that made
OverlayContainer focusable, the affordance widget became focusable which
caused some bugs. So, the affordance widget itself is now set as
non-focusable.

BUG=324516

Review URL: https://codereview.chromium.org/97483002