Fix flaky AppsSearchResultsControllerTest.UninstallReperformsSearch
Also fix the typo "reDperforms".

Since the asynchronous event that the test was waiting for is a Cocoa
event, the QuitClosure was only flakily ensuring that the event was
before it in the message queue. Using the same queue ensures the
messages are properly ordered.

BUG=308828

Review URL: https://codereview.chromium.org/98323002