Fix crash when a JSON cloud policy's value is invalid
When a JSON cloud policy is set to an invalid value, DecodeJson() returns
a NULL pointer, which causes the PolicyMap::Entry for this policy to have
a |value| member of NULL. This is unsupported. The |value| member must
always be non-NULL.

The CL fixes the crash by not adding a PolicyMap::Entry when the policy
value cannot be parsed.

BUG=None
TEST=Manual

Review URL: https://codereview.chromium.org/96903004