net: boost AES-GCM ciphers if the machine has AES-NI.
We want to use ChaCha20 based cipher suites on machines that don't have AES-NI
and AES-GCM on those that do.

Note: this was https://codereview.chromium.org/75663004/ and the review took
place there. But codereview broke and I had to reupload as a different issue.

BUG=310768
R=wtc@chromium.org

Review URL: https://codereview.chromium.org/91913002