Don't pull in content_switches.cc to whole chrome code base
The point of this change is remove compilation of content_switches.cc
that was done in installer_util target. This is a static library
that other parts of chrome pull in and that makes the Chrome code that
references those switches compile even if the switches itself are
not exported.
Installer_util doesn't even seem to uses those switches AFAIK.

Dropped compilation of these files so these problems are actually
exposed and addressed by the person adding the switch. The main
offender that referenced those switches was chrome's about_flags.cc.

This is based off rchlodnicki@opera.com's patch from https://codereview.chromium.org/24994002/

R=rchlodnicki@opera.com, robertshield@chromium.org

Review URL: https://codereview.chromium.org/85333008