events: Introduce EventTargeter.
EventTargeter will be used for finding the target for dispatching events in
aura and views. More details about the design at the document linked at the
bug.

This patch introduces an EventTargeter interface, with an empty default
implementation. It also adds some functions to the EventTarget interface
that will be used by the actual EventTargeter implementation.

BUG=318879
R=ben@chromium.org

Review URL: https://codereview.chromium.org/98293002