Start transitioning Pepper plugin code to using RenderFrame instead of RenderView.
This is part of the work in getting rid of RenderView. Since a lot of the code is tightly coupled, I'm added temporary getters on RenderFrame and RenderView to each other. This allows the work to be done in smaller steps.

This is also the proper fix for htt://crbug.com/275447.

BUG=275447, 245126
R=nasko@chromium.org

Review URL: https://codereview.chromium.org/93333002