Fix GN deadlock on error handling.
This would deadlock in some error cases when the failing thread tried to acquire the lock to report an error, but the main thread was in the process of shutting down (due to a previous error) and blocking on the worker threads finishing while holding thc lock.

This moves the worker pool shutdown to outside the lock, and adds some extra shutdown assertions.

It also fixes a syntax error in the base Linux build file.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/100193002