Make touch-based synthetic gesture take touch slop into account.
Touchscreen devices typically have a 'touch slop', i.e. a certain distance a
pointer has to move before it is considered moving. The touch slop is subtracted
from the distance a pointer moves. So if the touch slop is 10 pixels and a
pointer moves by 100 pixels on the screen, the resulting move is registered as
having covered only 90 pixels.

This patch adds the 'GetTouchSlopInDips' to SyntheticGestureTargets. This value
is used by the gestures to increase the distances covered by touch pointers. For
example, to scroll 100 pixels with a touch slop of 10 pixels, the pointer will
cover 110 pixels.

Touch-based synthetic gestures are updated to take the touch slop into account.

BUG=321114

Review URL: https://codereview.chromium.org/95153002