[Mac] Workaround for Mavericks AppKit accessibility bug in NSCollectionView.
Using the accessibility inspector on the Mac App Launcher in Mavericks
currently causes Chrome to crash. This is due to an AppKit bug that
triggers an unrecognized selector exception to be thrown on
-[NSButtonCell item], inside NSCollectionView code.

This adds a method to the NSButtonCell subclass used by the app launcher
for its NSCollectionView that provides a dummy implementation of the
method AppKit tries to call.

BUG=324365
TEST=Open Mac's Accessibility Inspector, then show the App Launcher.
Hovering over items in the grid should not crash Chrome.

Review URL: https://codereview.chromium.org/96103002