Add chrome://quit to shut down Chrome, including packaged apps.
Now that packaged apps are not quit when using wrench quit, there is no
way to completely shut down Chrome. While this is normally a good thing
as apps are intended to be separate from Chrome, for developers and
other power users, the ability to completely shut down Chrome can be a
useful function. This also blacklists chrome://quit and chrome://restart
from being saved to the user's session to prevent quit or restart loops.

BUG=317414

Review URL: https://codereview.chromium.org/84893002