[Mojo] Move JS bindings out of public
This CL moves the implementation of Mojo's JavaScript bindings out of the
"public" directory. The public directory is intended to be free of dependencies
on other parts of Chromium, but the JavaScript bindings depend on V8 and base.
Now, the bindings are just a subdirectory of mojo/apps/js, which is their only
client. This makes mojo.js more of a self-contained app.

R=darin@chromium.org

Review URL: https://codereview.chromium.org/100573003