Fix issues with token refresh in AccountReconcilor
The CL was change a bit as suggested by the reviewers.

This CL now introduces a worklist model in GoogleAutoLoginHelper. Instead of firing all the refresh token events in a row, it queues up all the login requests and insert GAIA cookies one-by-one. This would let accounts to be appended instead of overwriting one another.

This also introduces GoogleAutoLoginHelperTest for some unit testing.

BUG=305249

Review URL: https://codereview.chromium.org/63253003