Respond to WM_SETCURSOR when the cursor is in the client area
Because HWNDMessageHandler updates the cursor when in the non-client
area, it's overriding state set by CursorManger.

So, when Windows requests the cursor in the client area we need to
return the cursor that CursorManager set, otherwise there can't be any
caching up at that level.

R=sky@chromium.org
BUG=295556,312204

Review URL: https://codereview.chromium.org/96943004