Translate new UX: Fix the visual and behavior of 'Done' button
The label is 'Translate' only when the page is not translated in the current language pair, otherwise 'Done'.

When the label is 'Translate', the view state transits to 'Translating' from 'Advanced' after clicking the button, while when 'Done', this transits to the previous view state.

BUG=324696
TEST=unit_tests --gtest_filter=TranslateBubbleViewTest.*

Review URL: https://codereview.chromium.org/98643003