Sticky keys modifies an entire scroll sequence instead of just a single event.
A scroll sequence starts with a fling cancel, followed by a series of scroll
events and ends with a fling start event. In the oneshot sticky keys case, we
now modify the entire sequence with the modifier key before disabling sticky
keys.

Recommitting r236834 because it was reverted due to a memory leak.

BUG=308662
TEST=updated unit tests, ran valgrind test

Review URL: https://codereview.chromium.org/99213003