Remove DCHECK(context3d_) checks that check nothing.
We only allow creation of ContextProvider with a non-null context3d
and store it in the class inside the constructor. This was not always
the case, and we had DCHECKs to verify behaviour. Now, a DCHECK in
the constructor is enough (though also stricly not needed since the
Create() function checks for NULL). Remove the other DCHECKs as they
are just noise.

R=piman@chromium.org
BUG=

Review URL: https://codereview.chromium.org/102493003