Telemetry: Start scrolling at current position.
The scroll action would set scrollTop to 0 before scrolling, so we always start
from the top of the page. For pages that require multiple scrolls, e.g. because
more content is loaded dynamically when the bottom is reached, we want to scroll
from the current position rather than going back to the top.

In the typical case we only perform one scroll immediately after the page was
loaded, so this change shouldn't affect any measurements.

BUG=299828

Review URL: https://codereview.chromium.org/101323004