Expose GLES2Interface to cc
This exposes a gpu::gles2::GLES2Interface pointer to cc through the ContextProvider.  This is a way for the compositor to talk to the client side of the command buffer more directly, bypassing the blink::WebGraphicsContext3D indirection.  See the bug for the plan for using this new interface and deprecating WGC3D.

BUG=181120

Review URL: https://codereview.chromium.org/54463007