Unify WebGraphicsContext3D::insertSyncPoint impls
The in-process command buffer implementation of insertSyncPoint() does
a shallow flush and returns 0 as the sync point, since the flush posts a
task which ensures ordering. The out of process implementation does additional
work in the GpuControl implementation. This moves the logic to implement
the in process version into GLES2Implementation + GpuControl instead of short
circuiting in WebGraphicsContext3DInProcessCommandBufferImpl so that callers
can use the GLES2Interface version directly.

BUG=181120

Review URL: https://codereview.chromium.org/102813003