Remove gclient revert step from all recipe builders.
The gclient revert step was casing failures on the bots. This CL attempted to
remove the step: https://codereview.chromium.org/79613002/
The comments on that CL concluded that gclient revert is not needed. This CL
removes the step for all builds that use recipes.

BUG=313851

Review URL: https://codereview.chromium.org/89633002