Change ScreenshotTaker to use ScopedClipboardWriter.
WriteObjects() was originally designed so that it'd be possible to
serialize data from the renderer to be written to the clipboard in the
browser. Its implementation is going to become more sane, so refactor
callers that depend on the current quirky interface to use
ScopedClipboardWriter instead.

BUG=319285

Review URL: https://codereview.chromium.org/103233002