gin::Wrappable shouldn't inherit from base::RefCounted
Rather than using reference counting to keep gin::Wrappable objects alive, we
use the V8 garbage collector. To retain a reference to a wrappable object on
the stack, use a gin::Handle. Currently there is no mechanism for retaining a
gin::Wrappable object in the heap, but we'll likely add one in the future.

R=aa@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/105423003