Add a JNI AudioRecord implementation of AudioInputStream.
Add a native AudioRecordInputStream class which wraps a Java
AudioRecordInput class doing most of the work. This code will not yet
be used as it isn't hooked up in AudioManagerAndroid.

The purpose of this is to support enabling hardware input effects such
as echo cancellation and noise suppression. Although the OpenSL layer
has an interface for this, it's not currently implemented in Android.
We must go through an AudioRecord path to enable them.

On a Nexus 5, I measured a loopback apprtc latency of 190 - 195 ms.
The existing OpenSLES input measured at 205 - 215 ms.

TESTED=With audio_manager_android.cc modified to create
AudioRecordInputStream rather than OpenSLESInputStream:
- "media_unittests --gtest-filter=AudioAndroid*" passes on a Nexus 5.
- loopback apprtc call on Nexus 7 and Nexus 5 sounds good.
- git try succeeds.

Review URL: https://codereview.chromium.org/89983003