chromium-status: automatically linkify usernames/status messages
Automatically convert all usernames in the Who col to mailto links.
This makes the field much narrower w/out reducing information which
allows more space for the message field.

There is a large amount of links that can be created based on the status
messages, both from the automatic stuff the bots produce as well as the
common conventions of developers.  Add support for declaring regexes that
in turn create links of the data.

We also add support for doing instance-specific linking (for cases where
source data is ambiguous as to the target, but make sense when considered
in the instance they are shown).  We only use this for ChromeOS atm, but
should be easy to extend if people are interested.

Hopefully this should help out sheriffs by cutting out a few steps.

BUG=chromium:314501
TEST=ran an appengine instance with example statuses and checked output
TEST=`./tests/main_test.py` passes

Review URL: https://codereview.chromium.org/84943003