Add UMA stats for discarded tabs on a per logged in user base
Each discarded tab will be "stored" in the bucket for the number of users logged in at the time of the occurrence. 
To get to a useful number, the value needs to be normalized by dividing the absolute value per bucket by the corresponding UsersPerSession bucket (in percent).

BUG=321739
TEST=visual

Review URL: https://codereview.chromium.org/104713003