Changed the precache config settings proto to not include the whitelist of starting URLs.
The precache fetcher will now attempt to fetch manifests for the first |top_sites_count| starting URLs. Fetches of manifests for non-whitelisted URLs will fail and be skipped.

The reason for this change is so that the server side component can have a large whitelist of starting URLs, without forcing the client to download the entire list every time precaching is done.

BUG=306185

Review URL: https://codereview.chromium.org/83283004