Fix telemetry reconnect when lost to inspector
Telemetry used to rely on the websocket listing going away on the browser
when the inspector was gone. Now, it is there but the connect attempt
raises an error. This is the narrowest fix I can think of: we still will
fail on reconnect when the exception we get back is different. Broadening
that test can be done iff we see that this is overspecific.

BUG=324604

Review URL: https://codereview.chromium.org/96343003