Wait for Preference Syncing To Start Before Checking Welcome Notification Preferences
Starting up the Chrome Now Extension and Preference Syncing occur at the same
time. As a result, there is a race condition between checking the preferences
to show the welcome notification and syncing the preferences that determine
if we should show the welcome notification.

On Chrome runs with no existing profile data, when you sign in with a user
that previously saw the welcome notification, by the time Chrome Now displayed
its first card, the preferences weren't yet synced. This resulted in showing
the welcome notification.

The preference sync marks the welcome notification as shown
(kWelcomeNotificationDismissed), and the preference listener promptly hides
the welcome notification. This results in a welcome notification that is
shown and then quickly hidden away.

This change adds some waiting logic to the Welcome Notification. In short,
if the preferences are not yet syncing, wait until they are before checking
the dismissal preference.

BUG=318400

Review URL: https://codereview.chromium.org/100373003