Add AppListViewDelegateObserver to the app list, for profile changes.
The app list UI code is currently not receiving notifications for 
changes to set of profiles that are currently available in Chrome. 
E.g., when a profile is added or removed. This means the app list menu 
does not update when it should.

This CL removes signin status from the app list model, and adds an 
observer to the app list view delegate to handle signin status 
changes, as well as changes to the list of profiles available. These 
are intertwined, because a signin change results in the menu changing 
between the (signed in) email address, and the profile name.

BUG=320409
TEST=Add or remove a profile from Chrome. The App launcher menu 
should immediately reflect the new set of profiles.

Review URL: https://codereview.chromium.org/104893004