Zip utf-8 file names with encoding bit set.
This commit sets the 'locale' bit in global directory of a zip which
means that the file name was written in UTF-8. Chromium always writes
and reads zips as UTF-8 so it is not a problem per se but other, more
zip compliant, programs see packed files with wrong paths if this bit
is not set.

BUG=325737

Review URL: https://codereview.chromium.org/94093003