aura: Create a new FrameProvider when frame size in DIP changes.
When the renderer gives the UI a new frame that has the same number
of physical pixels but a new scale factor, we must display the frame
at a new scale in the UI.

In order to do this, we replace the FrameProvider entirely to use the
same path that we would use when the frame changed size (ie number of
pixels). This should help ensure that the new scale given to the ui::Layer
will be shown in sync with the new frame texture regardless of future
commit-less update type optimizations.

Test: RenderWidgetHostViewAuraTest.SoftwareDPIChange

R=piman
BUG=318410

Review URL: https://codereview.chromium.org/107143003