Remove direct-to-backbuffer Ganesh path
The usage of the GrContext in GLRenderer is getting in the way of
cleaning up the WebGraphicsContext3D usage in cc.

This path was good for trying out the Ganesh path, but was never really
used, available on all platforms, or compatible with ubercompositor.
Ganesh-to-tile is almost landed and so it seems like this can be safely
removed for the present.

BUG=none
TBR=sky

Review URL: https://codereview.chromium.org/98463008