Multiple level undo/redo of bookmarks available through Bookmark Bar.
The BookmarkUndoService is now being constructed and is observing
user initiated changes to the bookmark model.

Undo and Redo of bookmark changes can be triggered from the
Bookmark Bar context menu. This functionality is behind the switch
enable-bookmark-undo.

Still outstanding is the multi-level undo to be accessible from
the Bookmark Manager.

BUG=126092

Review URL: https://codereview.chromium.org/70233015