cc: Add histogram for unserviced copy requests and return by reference.
The LayerTreeImpl::LayersWithCopyOutputRequest() method is returning
a const std::vector<> instead of a const std::vector<>&. Oops.

But for this to be a huge perf problem, we must have a large vector
somehow. So adding a UMA histogram to tell us if the browser
is doing awful things and requesting a whole lot of copy requests
on undrawable subtrees.

R=enne
BUG=326617

Review URL: https://codereview.chromium.org/108853004