Add a PDF component extension
This adds a PDF component extension which is used for rendering PDFs with an
out of process PDF plugin. It adds the extension as a MIME handler for pdf file
types. The extension is currently only added if the --out-of-process-pdf switch
is passed to Chrome, AND the PDF plugin is available.

BUG=303491

Review URL: https://codereview.chromium.org/96233002