Don't unmap ports in host_driven/test_case.py.
The sync tests failure seems to have to do with port unmapping. Not unmapping
ports let most of the sync tests pass (only 2 failures remain for an unrelated
reason apparently).
Note that port unmapping was added to avoid exceeding the file descriptor limit
in net_unittests since each test server was listening on a different port. The
sync tests use a server always listening on a same port so unmapping the port
is not strictly needed.

BUG=313809

Review URL: https://codereview.chromium.org/106533002