When calculating the minimum toolbar size in popup mode, use the location bar's preferred width, just as we do in non-popup mode.
Technically this also sets the minimum height in non-popup mode to the larger of
the location bar height and the background image height, but that should never
have an effect right now.  It will probably be necessary for a vertically-auto-
expanding omnibox down the road, though.

BUG=none
TEST=none
R=msw@chromium.org

Review URL: https://codereview.chromium.org/103023003