Revert/Fix "Disable PnaclHeaderTest.TestHasPnaclHeader."
This reverts r236737.

The test was failing consistently because of a DCHECK added
in 236651, which was then reverted in 236693, but re-applied
at 236805 after this test was disabled. The issue is that
this test uses EmbeddedTestServer in a way that violated the
DCHECK (http://crbug.com/321422). We have to wait till after
SetUp to start the server now (perhaps the example usage
comments in embedded_test_server.h should no longer recommend
calling InitializeAndWaitUntilReady() in SetUp?).

Also, the dtor automatically does ShutdownAndWaitUntilComplete,
so no longer need to put that under TearDown().

BUG=315328
R=ncbray@chromium.org, oshima@chromium.org

Review URL: https://codereview.chromium.org/83073010