Hide scrollbars in overscroll tests.
The overscroll tests in WebContentsViewAuraTest are currently relying
on broken behaviour in the scrollbar. Some touches in the test are
originating over the scrollbars. Scrollbars shouldn't have any impact
on these tests, so we should hide them.

This was the cause of failures due to this patch:
https://codereview.chromium.org/102663003/

TEST=WebContentsViewAuraTest.*

Review URL: https://codereview.chromium.org/110403002