Disable the 'Manage exceptions' button for protected content.
The content setting for protected content identifiers has exceptions
which record user response to notifications.  These responses (in
addition to manually added exceptions) are only considered when the
option is enabled.  If a user disables this option then identifiers will
be disabled for all sites.  To reflect these semantics, the 'Manage
exceptions' button is now disabled when the option is disabled.

BUG=chromium:318902
TEST=manual

Review URL: https://codereview.chromium.org/108623003